Worked on benchmark utils
